:white_check_mark: Add tests for client config loading
